Use relative mouse mode
This addresses the issue of the mouse pointer being out of sync in and
out of the emulator.

Now left click locks the mouse and hides the system mouse pointer. Only
in this state, the PC-98 side will get motion and button events. To
unlock the cursor, press the middle button to display the menu.