Use SDL_Scancode rather than SDL_Keycode
Before this change, keyboard events were handled using SDL_Keycode,
which meant that some keys could not be used, if the user was using a
non-JP keyboard.

After this, key handling uses SDL_Scancode. This makes more sense
because the emulator is interested in the physical position of the keys,
not the characters. Even if the user is using a US keyboard, the
emulator will work with the PC-98 layout.

The mapping of "international" keys follows the suggestion in the USB
HID Usage Tables standard (page 60 of [1]).

[1] https://www.usb.org/sites/default/files/documents/hut1_12v2.pdf