Show mouse cursor when pointer lock is deactivated
The browser releases pointer lock for various reasons (e.g., when ESC
is pressed or browser tabs are switched). SDL tries to behave as if the
mouse is still captured, but that is confusing because the cursor does
not appear when the user press the ESC key.

So, this change listens to Emscripten's pointerlockchange event and
releases mouse capture when the pointer is unlocked.